Add Component Scan and fix package typo
