+++ src/main/java/com/springframework/pet/services/CatService.java
@@ -1,0 +1,13 @@
+package com.springframework.pet.services;
+
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Service;
+
+@Profile("cat")
+@Service
+public class CatService implements PetService {
+    @Override
+    public String getPets() {
+        return "Cats";
+    }
+}
