+++ src/main/java/com/springframework/pet/services/DogService.java
@@ -1,0 +1,13 @@
+package com.springframework.pet.services;
+
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Service;
+
+@Profile({"dog", "default"})
+@Service
+public class DogService implements PetService {
+    @Override
+    public String getPets() {
+        return "Dogs";
+    }
+}
