+++ src/main/java/com/springframework/pet/services/PetService.java
@@ -1,0 +1,5 @@
+package com.springframework.pet.services;
+
+public interface PetService {
+    String getPets();
+}
