+++ src/main/java/guru/springframework/sfgdi/controllers/ConstructorInjectedController.java
@@ -1,0 +1,19 @@
+package guru.springframework.sfgdi.controllers;
+
+import guru.springframework.sfgdi.services.GreetingService;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class ConstructorInjectedController {
+    private final GreetingService greetingService;
+
+    // no need to @Autowired on a constructor
+    public ConstructorInjectedController(@Qualifier("constructorInjectedGreetingService") GreetingService greetingService) {
+        this.greetingService = greetingService;
+    }
+
+    public String getGreeting(){
+        return greetingService.sayGreeting();
+    }
+}
