+++ src/main/java/guru/springframework/sfgdi/controllers/MyController.java
@@ -1,0 +1,19 @@
+package guru.springframework.sfgdi.controllers;
+
+import guru.springframework.sfgdi.services.GreetingService;
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class MyController {
+
+    private final GreetingService greetingService;
+
+    public MyController(GreetingService greetingService) {
+        this.greetingService = greetingService;
+    }
+
+    public String sayHello() {
+
+        return greetingService.sayGreeting();
+    }
+}
