+++ src/main/java/guru/springframework/sfgdi/controllers/PropertyInjectedController.java
@@ -1,0 +1,17 @@
+package guru.springframework.sfgdi.controllers;
+
+import guru.springframework.sfgdi.services.GreetingService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class PropertyInjectedController {
+    @Qualifier("propertyInjectedGreetingService")
+    @Autowired
+    public GreetingService greetingService;
+
+    public String getGreeting() {
+        return greetingService.sayGreeting();
+    }
+}
