+++ src/main/java/guru/springframework/sfgdi/controllers/SetterInjectedController.java
@@ -1,0 +1,20 @@
+package guru.springframework.sfgdi.controllers;
+
+import guru.springframework.sfgdi.services.GreetingService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class SetterInjectedController {
+    private GreetingService greetingService;
+
+    @Autowired
+    public void setGreetingService(@Qualifier("setterInjectedGreetingService") GreetingService greetingService) {
+        this.greetingService = greetingService;
+    }
+
+    public String getGreeting(){
+        return greetingService.sayGreeting();
+    }
+}
