+++ src/main/java/guru/springframework/sfgdi/services/ConstructorInjectedGreetingService.java
@@ -1,0 +1,11 @@
+package guru.springframework.sfgdi.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ConstructorInjectedGreetingService implements GreetingService {
+    @Override
+    public String sayGreeting() {
+        return "Hello World! -- Constructor";
+    }
+}
