+++ src/main/java/guru/springframework/sfgdi/services/GreetingService.java
@@ -1,0 +1,5 @@
+package guru.springframework.sfgdi.services;
+
+public interface GreetingService {
+    public String sayGreeting();
+}
