+++ src/main/java/guru/springframework/sfgdi/services/GreetingServiceImpl.java
@@ -1,0 +1,13 @@
+package guru.springframework.sfgdi.services;
+
+import org.springframework.context.annotation.Primary;
+import org.springframework.stereotype.Service;
+
+@Primary
+@Service
+public class GreetingServiceImpl implements GreetingService {
+    @Override
+    public String sayGreeting() {
+        return "Hello World! -- Primary";
+    }
+}
