+++ src/main/java/guru/springframework/sfgdi/services/I18nEnglishGreetingService.java
@@ -1,0 +1,13 @@
+package guru.springframework.sfgdi.services;
+
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Service;
+
+@Profile({"EN", "default"})
+@Service("i18nService")
+public class I18nEnglishGreetingService implements GreetingService{
+    @Override
+    public String sayGreeting() {
+        return "Hello World - EN";
+    }
+}
