+++ src/main/java/guru/springframework/sfgdi/services/PropertyInjectedGreetingService.java
@@ -1,0 +1,11 @@
+package guru.springframework.sfgdi.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class PropertyInjectedGreetingService implements GreetingService {
+    @Override
+    public String sayGreeting() {
+        return "Hello World! -- Property";
+    }
+}
