+++ src/main/java/guru/springframework/sfgdi/services/SetterInjectedGreetingService.java
@@ -1,0 +1,11 @@
+package guru.springframework.sfgdi.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class SetterInjectedGreetingService implements GreetingService {
+    @Override
+    public String sayGreeting() {
+        return "Hello World! -- Setter";
+    }
+}
