--- src/main/java/guru/springframwork/sfgdi/controllers/ConstructorInjectedController.java
@@ -1,19 +1,0 @@
-package guru.springframwork.sfgdi.controllers;
-
-import guru.springframwork.sfgdi.services.GreetingService;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class ConstructorInjectedController {
-    private final GreetingService greetingService;
-
-    // no need to @Autowired on a constructor
-    public ConstructorInjectedController(@Qualifier("constructorInjectedGreetingService") GreetingService greetingService) {
-        this.greetingService = greetingService;
-    }
-
-    public String getGreeting(){
-        return greetingService.sayGreeting();
-    }
-}
