--- src/main/java/guru/springframwork/sfgdi/controllers/I18nGreetingController.java
@@ -1,18 +1,0 @@
-package guru.springframwork.sfgdi.controllers;
-
-import guru.springframwork.sfgdi.services.GreetingService;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class I18nGreetingController {
-    private final GreetingService greetingService;
-
-    public I18nGreetingController(@Qualifier("i18nService") GreetingService greetingService) {
-        this.greetingService = greetingService;
-    }
-
-    public String getGreeting() {
-        return greetingService.sayGreeting();
-    }
-}
