--- src/main/java/guru/springframwork/sfgdi/controllers/MyController.java
@@ -1,19 +1,0 @@
-package guru.springframwork.sfgdi.controllers;
-
-import guru.springframwork.sfgdi.services.GreetingService;
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class MyController {
-
-    private final GreetingService greetingService;
-
-    public MyController(GreetingService greetingService) {
-        this.greetingService = greetingService;
-    }
-
-    public String sayHello() {
-
-        return greetingService.sayGreeting();
-    }
-}
