--- src/main/java/guru/springframwork/sfgdi/controllers/PropertyInjectedController.java
@@ -1,17 +1,0 @@
-package guru.springframwork.sfgdi.controllers;
-
-import guru.springframwork.sfgdi.services.GreetingService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class PropertyInjectedController {
-    @Qualifier("propertyInjectedGreetingService")
-    @Autowired
-    public GreetingService greetingService;
-
-    public String getGreeting() {
-        return greetingService.sayGreeting();
-    }
-}
