--- src/main/java/guru/springframwork/sfgdi/controllers/SetterInjectedController.java
@@ -1,20 +1,0 @@
-package guru.springframwork.sfgdi.controllers;
-
-import guru.springframwork.sfgdi.services.GreetingService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class SetterInjectedController {
-    private GreetingService greetingService;
-
-    @Autowired
-    public void setGreetingService(@Qualifier("setterInjectedGreetingService") GreetingService greetingService) {
-        this.greetingService = greetingService;
-    }
-
-    public String getGreeting(){
-        return greetingService.sayGreeting();
-    }
-}
