--- src/main/java/guru/springframwork/sfgdi/services/ConstructorInjectedGreetingService.java
@@ -1,11 +1,0 @@
-package guru.springframwork.sfgdi.services;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class ConstructorInjectedGreetingService implements GreetingService {
-    @Override
-    public String sayGreeting() {
-        return "Hello World! -- Constructor";
-    }
-}
