--- src/main/java/guru/springframwork/sfgdi/services/GreetingService.java
@@ -1,5 +1,0 @@
-package guru.springframwork.sfgdi.services;
-
-public interface GreetingService {
-    public String sayGreeting();
-}
