--- src/main/java/guru/springframwork/sfgdi/services/GreetingServiceImpl.java
@@ -1,13 +1,0 @@
-package guru.springframwork.sfgdi.services;
-
-import org.springframework.context.annotation.Primary;
-import org.springframework.stereotype.Service;
-
-@Primary
-@Service
-public class GreetingServiceImpl implements GreetingService {
-    @Override
-    public String sayGreeting() {
-        return "Hello World! -- Primary";
-    }
-}
