--- src/main/java/guru/springframwork/sfgdi/services/I18nEnglishGreetingService.java
@@ -1,13 +1,0 @@
-package guru.springframwork.sfgdi.services;
-
-import org.springframework.context.annotation.Profile;
-import org.springframework.stereotype.Service;
-
-@Profile({"EN", "default"})
-@Service("i18nService")
-public class I18nEnglishGreetingService implements GreetingService{
-    @Override
-    public String sayGreeting() {
-        return "Hello World - EN";
-    }
-}
