--- src/main/java/guru/springframwork/sfgdi/services/I18nSpanishGreetingService.java
@@ -1,13 +1,0 @@
-package guru.springframwork.sfgdi.services;
-
-import org.springframework.context.annotation.Profile;
-import org.springframework.stereotype.Service;
-
-@Profile("ES")
-@Service("i18nService")
-public class I18nSpanishGreetingService implements GreetingService{
-    @Override
-    public String sayGreeting() {
-        return "Hola Mundo - ES";
-    }
-}
