--- src/main/java/guru/springframwork/sfgdi/services/PropertyInjectedGreetingService.java
@@ -1,11 +1,0 @@
-package guru.springframwork.sfgdi.services;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class PropertyInjectedGreetingService implements GreetingService {
-    @Override
-    public String sayGreeting() {
-        return "Hello World! -- Property";
-    }
-}
