+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-#spring.profiles.active=ES
+spring.profiles.active=ES, cat
