+++ src/test/java/guru/springframework/sfgdi/SfgDiApplicationTests.java
@@ -1,4 +1,4 @@
-package guru.springframwork.sfgdi;
+package guru.springframework.sfgdi;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
