+++ src/test/java/guru/springframework/sfgdi/controllers/ConstructorInjectedControllerTest.java
@@ -1,0 +1,20 @@
+package guru.springframework.sfgdi.controllers;
+
+import guru.springframework.sfgdi.services.GreetingServiceImpl;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class ConstructorInjectedControllerTest {
+
+    ConstructorInjectedController controller;
+
+    @BeforeEach
+    void setUp() {
+        controller = new ConstructorInjectedController(new GreetingServiceImpl());
+    }
+
+    @Test
+    void getGreeting() {
+        System.out.println(controller.getGreeting());
+    }
+}
