+++ src/test/java/guru/springframework/sfgdi/controllers/PropertyInjectedControllerTest.java
@@ -1,0 +1,21 @@
+package guru.springframework.sfgdi.controllers;
+
+import guru.springframework.sfgdi.services.GreetingServiceImpl;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class PropertyInjectedControllerTest {
+
+    PropertyInjectedController controller;
+
+    @BeforeEach
+    void setUp() {
+        controller = new PropertyInjectedController();
+        controller.greetingService = new GreetingServiceImpl();
+    }
+
+    @Test
+    void getGreeting() {
+        System.out.println(controller.getGreeting());
+    }
+}
