+++ src/test/java/guru/springframework/sfgdi/controllers/SetterInjectedControllerTest.java
@@ -1,0 +1,21 @@
+package guru.springframework.sfgdi.controllers;
+
+import guru.springframework.sfgdi.services.GreetingServiceImpl;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class SetterInjectedControllerTest {
+
+    SetterInjectedController controller;
+
+    @BeforeEach
+    void setUp() {
+        controller = new SetterInjectedController();
+        controller.setGreetingService(new GreetingServiceImpl());
+    }
+
+    @Test
+    void getGreeting() {
+        System.out.println(controller.getGreeting());
+    }
+}
