--- src/test/java/guru/springframwork/sfgdi/controllers/ConstructorInjectedControllerTest.java
@@ -1,22 +1,0 @@
-package guru.springframwork.sfgdi.controllers;
-
-import guru.springframwork.sfgdi.services.GreetingServiceImpl;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class ConstructorInjectedControllerTest {
-
-    ConstructorInjectedController controller;
-
-    @BeforeEach
-    void setUp() {
-        controller = new ConstructorInjectedController(new GreetingServiceImpl());
-    }
-
-    @Test
-    void getGreeting() {
-        System.out.println(controller.getGreeting());
-    }
-}
