--- src/test/java/guru/springframwork/sfgdi/controllers/PropertyInjectedControllerTest.java
@@ -1,23 +1,0 @@
-package guru.springframwork.sfgdi.controllers;
-
-import guru.springframwork.sfgdi.services.GreetingServiceImpl;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class PropertyInjectedControllerTest {
-
-    PropertyInjectedController controller;
-
-    @BeforeEach
-    void setUp() {
-        controller = new PropertyInjectedController();
-        controller.greetingService = new GreetingServiceImpl();
-    }
-
-    @Test
-    void getGreeting() {
-        System.out.println(controller.getGreeting());
-    }
-}
