--- src/test/java/guru/springframwork/sfgdi/controllers/SetterInjectedControllerTest.java
@@ -1,23 +1,0 @@
-package guru.springframwork.sfgdi.controllers;
-
-import guru.springframwork.sfgdi.services.GreetingServiceImpl;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class SetterInjectedControllerTest {
-
-    SetterInjectedController controller;
-
-    @BeforeEach
-    void setUp() {
-        controller = new SetterInjectedController();
-        controller.setGreetingService(new GreetingServiceImpl());
-    }
-
-    @Test
-    void getGreeting() {
-        System.out.println(controller.getGreeting());
-    }
-}
